Release 2.3.1 (#221)
* Update README.zh-cn.md

* Delete README.2.2.md

* Fix  flush unclaer data bugs.

* Update README.md

* Update README.md

* Update README.zh-cn.md

* update readme.md to add the subscriber group descriptions.

* update readme

* update version to 2.3.1

* add SnowflakeId unit tests

* update build.cake for unit test adjustment

* remove unused config file

* update readme

* update readme

* Fixed dashboard messages requeue error. ( #205 )

* Refactoring

* Fixed SnowflakeId workerId to random id.

* Update README.md

* upgrade dependent nuget packages.

* Add SourceLink Support

+++ src/DotNetCore.CAP/Dashboard/BatchCommandDispatcher.cs
@@ -9,9 +9,9 @@ namespace DotNetCore.CAP.Dashboard
 {
     internal class BatchCommandDispatcher : IDashboardDispatcher
     {
-        private readonly Action<DashboardContext, int> _command;
+        private readonly Action<DashboardContext, long> _command;
 
-        public BatchCommandDispatcher(Action<DashboardContext, int> command)
+        public BatchCommandDispatcher(Action<DashboardContext, long> command)
         {
             _command = command;
         }
@@ -27,7 +27,7 @@ namespace DotNetCore.CAP.Dashboard
 
             foreach (var messageId in messageIds)
             {
-                var id = int.Parse(messageId);
+                var id = long.Parse(messageId);
                 _command(context, id);
             }
